python/services/agent_lda/requirements.txt
# Pin protobuf to a stable version compatible with google-cloud libraries
# protobuf==3.20.3
protobuf

# Web-Framework, um Pub/Sub-Nachrichten per HTTP-Push zu empfangen
fastapi
uvicorn[standard]

# Google Cloud-Bibliotheken (neueste stabile Versionen)
google-cloud-pubsub
google-cloud-firestore
google-cloud-logging
google-cloud-storage
google-api-core
google-cloud-monitoring
python-dotenv
